+++ src/main/java/core/di/beans/factory/ProxyFactoryBean.java
@@ -1,6 +1,6 @@
-package core.aop;
+package core.di.beans.factory;
 
-import core.di.beans.factory.FactoryBean;
+import core.aop.Advisor;
 import java.lang.reflect.Method;
 import java.util.stream.Stream;
 import net.sf.cglib.proxy.Enhancer;
